chore(project setup): setup app folder structure (#3)
[finishes #171426236]
