+++ src/pages/Homepage.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Landing from '../components/Landing';
+
+const Homepage = () => (
+    <Landing />
+);
+
+export default Homepage;
